Restructure UI concepts based on simplicity.
  * Show basic concepts with the built-in renderer in the root `ui` module.
  * Move runtime and customization concepts into a `ui::core` module.
  * Simplify example further.
