+++ src/ui/checkbox.rs
@@ -1,24 +1,27 @@
-use super::Basic;
+use super::Renderer;
 
 use crate::graphics::{Point, Rectangle, Sprite};
-use crate::ui::{radio, MouseCursor};
+use crate::ui::core::{widget::checkbox, MouseCursor};
+
+pub type Checkbox<M> = checkbox::Checkbox<M, Renderer>;
 
 const SPRITE: Rectangle<u16> = Rectangle {
     x: 98,
-    y: 28,
+    y: 0,
     width: 28,
     height: 28,
 };
 
-impl radio::Renderer for Basic {
+impl checkbox::Renderer for Renderer {
     fn draw(
         &mut self,
-        is_selected: bool,
+        is_checked: bool,
         bounds: Rectangle<f32>,
-        bounds_with_label: Rectangle<f32>,
+        text_bounds: Rectangle<f32>,
         cursor_position: Point,
     ) -> MouseCursor {
-        let mouse_over = bounds_with_label.contains(cursor_position);
+        let mouse_over = bounds.contains(cursor_position)
+            || text_bounds.contains(cursor_position);
 
         self.sprites.add(Sprite {
             source: Rectangle {
@@ -29,7 +32,7 @@ impl radio::Renderer for Basic {
             scale: (1.0, 1.0),
         });
 
-        if is_selected {
+        if is_checked {
             self.sprites.add(Sprite {
                 source: Rectangle {
                     x: SPRITE.x + SPRITE.width * 2,
