+++ src/ui/core/hasher.rs
